requirements/base.txt
PyJWT==1.4.0
websockets==2.6
